requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
drf-yasg
redis
django-redis
celery
